Refact controllers with service
