+++ app/controllers/auth/auth_facebook_controller.ts
@@ -1,8 +1,6 @@
-import type { HttpContext } from '@adonisjs/core/http'
-import env from '#start/env'
-import User from '#models/user'
-import { AuthProviders } from '../../enums/auth-providers.enum.js'
-import { isUniqueProvider } from '../auth/guards/auth_provider_guard.js'
+// Importez le nouveau service
+import SocialAuthService from '#services/social_auth_service'
+import { HttpContext } from '@adonisjs/core/http'
 
 export default class AuthFacebooksController {
   async redirect({ ally }: HttpContext) {
@@ -14,48 +12,16 @@ export default class AuthFacebooksController {
   async handleCallback({ ally, response }: HttpContext) {
     const facebookUser = ally.use('facebook')
 
-    if (facebookUser.accessDenied()) {
-      return 'Access Denied'
+    // Utilisez le service pour gérer le callback de Facebook
+    const result = await SocialAuthService.handleFacebookCallback(facebookUser)
+
+    if (typeof result === 'string') {
+      return response.badRequest({ message: result })
     }
-
-    if (facebookUser.stateMisMatch()) {
-      return 'Request expired. Retry again'
-    }
-
-    if (facebookUser.hasError()) {
-      return facebookUser.getError()
-    }
-
-    const user = await facebookUser.user()
-
-    const isProviderVerified = await isUniqueProvider(user.email, AuthProviders.facebook)
-
-    if (!isProviderVerified) {
-      return response
-        .status(400)
-        .send({ message: `Already exists on other provider: ${AuthProviders.facebook}` })
-    }
-
-    const findUser = {
-      email: user.email as string,
-    }
-
-    const userDetails = {
-      email: user.email,
-      fullName: user.original.first_name,
-      provider: AuthProviders.facebook as string,
-      providerId: user.id,
-    }
-
-    const newUser = await User.firstOrCreate(findUser, userDetails)
-
-    const token = await User.accessTokens.create(newUser, ['*'], {
-      expiresIn: env.get('JWT_EXPIRY'),
-    })
 
     return response.ok({
-      token: token,
-      ...newUser.serialize(),
+      token: result.token,
+      ...result.user.serialize(),
     })
   }
 }
