+++ app/controllers/auth/auth_google_controller.ts
@@ -1,9 +1,5 @@
-import env from '#start/env'
 import type { HttpContext } from '@adonisjs/core/http'
-import User from '#models/user'
-import { AuthProviders } from '../../enums/auth-providers.enum.js'
-
-import { isUniqueProvider } from '../auth/guards/auth_provider_guard.js'
+import SocialAuthService from '#services/social_auth_service'
 
 export default class AuthGoogleController {
   async redirect({ ally }: HttpContext) {
@@ -13,48 +9,16 @@ export default class AuthGoogleController {
   async handleCallback({ ally, response }: HttpContext) {
     const googleUser = ally.use('google')
 
-    if (googleUser.accessDenied()) {
-      return 'Access Denied'
+    // Utilisez le service pour gérer le callback de Google
+    const result = await SocialAuthService.handleGoogleCallback(googleUser)
+
+    if (typeof result === 'string') {
+      return response.badRequest({ message: result })
     }
-
-    if (googleUser.stateMisMatch()) {
-      return 'Request expired. Retry again'
-    }
-
-    if (googleUser.hasError()) {
-      return googleUser.getError()
-    }
-
-    const user = await googleUser.user()
-
-    const isProviderVerified = await isUniqueProvider(user.email, AuthProviders.google)
-
-    if (!isProviderVerified) {
-      return response
-        .status(400)
-        .send({ message: `Already exists on other provider: ${AuthProviders.google}` })
-    }
-
-    const findUser = {
-      email: user.email as string,
-    }
-
-    const userDetails = {
-      email: user.email,
-      fullName: user.original.given_name,
-      provider: AuthProviders.google as string,
-      providerId: user.id,
-    }
-
-    const newUser = await User.firstOrCreate(findUser, userDetails)
-
-    const token = await User.accessTokens.create(newUser, ['*'], {
-      expiresIn: env.get('JWT_EXPIRY'),
-    })
 
     return response.ok({
-      token: token,
-      ...newUser.serialize(),
+      token: result.token,
+      ...result.user.serialize(),
     })
   }
 }
